fix(billing): preserve portal access through runtime overrides
